test(rust): expand core coverage and separate unit test modules
- cover metric saturation and response budget ownership
- verify reservation cleanup when decompression is bypassed
- move inline Rust tests into dedicated adjacent modules
- document Rust test organization conventions
